+++ packages/scripts/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Feature
 
+- Add `--alpha` option to use latest alpha in generator @sneridagh
+
 ### Bugfix
 
 ### Internal